login: Document environment variables with the "envvar" directive
This makes them linkable/referencable (e.g. with the "envvar" role) and
indexed in the web docs and resolves warnings in the Sphinx build about
missing reference targets.  Since "envvar" is a Sphinx directive, not a
core docutils/rST directive, I've implemented a basic version of it for
our rst_to_text() function used for --help output.