runner.singularity: Clean up caches during update and setup
Since we immediately convert the Docker/OCI image to SIF, the Docker/OCI
image layer caches end up being duplicative.  Not keeping them cached
means the cache can't be used to avoid downloading layers that haven't
changed in subsequent updates, but I think when choosing between
minimizing transfer or storage for our users, we should minimize
storage.

If we want to have our cake and eat it to (i.e. minimize both), we might
be able to change tack to newer Singularity features where (I think) OCI
images can be used directly without first converting to SIF.  Then, like
with our Docker runtime, the image we use and layer caches are one and
the same.  That's a larger change for a future time, though.