Add UsageError for when we want to print command usage before the error message
Normally argparse does this when it detects a usage error, but this lets
us handle other cases it doesn't in a similar way.  Similar to how we
track which command class the args parse to, we track the command parser
the args parse to so we can emit appropriate usage information.

For consistency with argparse, the exit status of a usage error is 2
instead of 1, but the distinction is unlikely to be actually useful in
practice.