dev: Fix new errors in type checking around the type of Ellipsis (...)
pyright 1.1.325 (released yesterday) added support for @type_check_only
decorators¹, which were added (a year ago) to the builtins.ellipsis type
of the stdlib typeshed².  This type is available only during
type-checking and doesn't actually exist at runtime, but previously
pyright didn't care about that (and mypy still doesn't).  Now it does,
noticed thanks to scheduled CI, and throws an error like:

    "ellipsis" is marked as @type_check_only and can be used only in
    type annotations (reportGeneralTypeIssues)

Explicitly clarify to pyright that our variable is a type alias³ and not
a value also used at runtime so it knows its ok to use the type of
Ellipsis (via a forward reference to the now-you-see-it, now-you-don't
builtins.ellipsis).

Relatedly, we can switch from 'builtins.ellipsis' → types.EllipsisType
on Python ≥3.10⁴ and stop using the private/internal type for Ellipsis.
I'll do that in the subsequent commit.

¹ <https://github.com/microsoft/pyright/pull/5817>

² <https://github.com/python/typeshed/commit/a92da58328aa259184ff45bb7408e82426fb563e>
  <https://github.com/python/typeshed/pull/8531>

³ <https://docs.python.org/3/library/typing.html#typing.TypeAlias>
  <https://peps.python.org/pep-0613/>

⁴ <https://docs.python.org/3/library/types.html#types.EllipsisType>
  <https://github.com/python/cpython/issues/85976>