build: Add --detach-on-interrupt for AWS Batch jobs
Its sometimes useful to treat Ctrl-C/SIGINT as non-terminating for the
remote job, for example in automation contexts (e.g. GitHub Actions)
where SIGINT may be sent as part of the automation service's job control
system or even merely as a molly-guard to prevent accidental
cancellation when attaching to a build to only observe it.