runner.aws_batch: Detach on SIGHUP too, when present
SIGHUP happens, for example, if the user closes a terminal (or the
terminal dies on its own) while our process is still running.  It can
happen in other cases too.  The remote job stays running whether we
handle SIGHUP or not, but it's nice to detach cleanly and be explicit
about our expectations.