dev: Test against the SingularityCE 3.x series in CI
The release of SingularityCE 4.0.0 a few hours ago¹ caused a failure
during CI for the release of Nextstrain CLI 7.3.0.²  We got unlucky as
that release's CI was the first to encounter the issue due to timing.
We can test the compatibility of SingularityCE 4.x with both our CI
setup process and Singularity runtime at some later point.

¹ <https://github.com/sylabs/singularity/releases/tag/v4.0.0>
² <https://github.com/nextstrain/cli/actions/runs/6238951959>