aws.cognito: Update the refresh token when renewing tokens, if applicable
Reading the OAuth 2.0 spec again¹, I noted that:

    The authorization server MAY issue a new refresh token, in which
    case the client MUST discard the old refresh token and replace it
    with the new refresh token. The authorization server MAY revoke the
    old refresh token after issuing a new refresh token to the client.

I had assumed the refresh token itself was never renewed as I never
observed AWS Cognito doing so in practice.  I addressed this assumption
for nextstrain.org², which is using OAuth 2.0, but it also got me
thinking about Nextstrain CLI here, which isn't using OAuth 2.0 per se
but is using a Cognito API that's very similar.

Like the OAuth 2.0 spec, the documented Cognito response type
definitions include the possibility of an updated refresh token.³
Perhaps Cognito even does renew them, but only as necessary and we've
never happened to notice this issue!

Regardless, we're planning to move to generic OAuth 2.0 authentication
instead of using Cognito's API, or at least support the former in
addition to the latter.  This change of logic will help smooth that
switch.

¹ <https://datatracker.ietf.org/doc/html/rfc6749#section-6>
² <https://github.com/nextstrain/nextstrain.org/pull/736>
³ <https://docs.aws.amazon.com/cognito-user-identity-pools/latest/APIReference/API_AuthenticationResultType.html>