Remove TODO on using parse_intermixed_args
The motivation to use argparse was presumably to replace the hardcoded
prefix matching with argparse's built-in functionality¹. That seems
tricky since the current prefix matching omits some patterns (e.g. --co
for --cores, -re for --resources) which would be included by argparse
unless some workarounds are applied.

The current implementation seems reasonable, so leave it as-is.

¹ https://docs.python.org/3.8/library/argparse.html#prefix-matching