runner.singularity: Explicitly support Apptainer's `singularity` when checking versions
My intent was (and remains) to try to maintain compatibility with both
SingularityCE and Apptainer after the Singularity fork.  I didn't
realize/recall that Apptainer restarted version numbering when I added
this version inspection!

It would be most ideal if Apptainer's `singularity` executable reported
the equivalent/compatible SingularityCE version, but that's
understandably difficult and so I get why it reports the Apptainer
version instead.

Thanks to @osageorange for raising this issue, testing this patch, and
reporting back that it works!

As SingularityCE and Apptainer continue to diverge, a more complicated
treatment of them may be needed, but for now this should suffice.

Resolves <https://github.com/nextstrain/cli/issues/329>.