remote.nextstrain_dot_org: Head off potential bugs in authn error handling
There was a narrow space for bugs here since it was reconstructing what
the user associated with the request was *expected* to be, but not what
it necessarily *actually* was, e.g. what was used to send the actual

    Authorization: Bearer …

header in the request.

From a given requests.Response object we can't access the auth provider
directly (by design), so we have to arrange a backchannel to pass along
to the error handler the actual user authn information used in the
request.  A simple custom attribute on the request object does the job
nicely.

One alternative would be a custom request header, e.g.

    Nextstrain-Username: trs

or some such, but I like the attribute because it's more targeted in
scope and doesn't send unnecessary information.  Another alternative
would be for the error handler to parse the token in the Authorization
header—it doesn't really even need to verify the JWT, just parse it—and
extract the username that way.  But that's a bit more involved and the
simplicity of a custom attribute is more compelling.