dev: Appease Pyright 1.1.345's changed type checking of tuple slices
A slice like x[0:2] may evaluate to ('a',) or ('a', 'b',) depending on
the length of x.  Pyright changed its checking of this slice behaviour
for tuples in 1.1.345 (released 4 days ago)¹, which started raising
issues with this code in CI.

While we humans know that the sockaddr tuples here always have at least
two elements, Pyright does not.  I think Pyright could do better here as
it knows the type of sockaddr is

    Tuple[str, int] | Tuple[str, int, int, int]

and thus I think should be able to know for sure the type of a 0:2
slice.  Instead, it evaluates the slice type to an ambiguous

    Tuple[str | int, ...]

which is incompatible with our declared return type.

I went to report this behaviour regression, but someone's already beat
me to it.²

¹ <https://github.com/microsoft/pyright/releases/tag/1.1.345>
² <https://github.com/microsoft/pyright/issues/6958>