authn.configuration: Improve error handling in openid_configuration()
@victorlin suggested¹ we improve error handling in this code path, and
he's right.

Catch and improve the output for the three most common errors we'll
encounter: connection errors, HTTP errors, and JSON decoding errors.
The original error is summarized in the error output, and by chaining
the new exception from the original, the full chain in all its detail
will be printed when NEXTSTRAIN_DEBUG=1 for troubleshooting.

¹ <https://github.com/nextstrain/cli/pull/333#discussion_r1468122701>