standalone-installer-unix: Use new aarch64 builds for macOS
Conditions on the Nextstrain CLI version being requested, as aarch64
builds¹ will only be available going forward for new releases.  Assumes
the next release will be 8.2.0!

Disables Rosetta check when using an aarch64 build, as the checks will
be done at the runtime-level instead by `nextstrain setup` and
`nextstrain check-setup`.²

We must wait to merge this until after the first release (i.e. 8.2.0) of
the changes it depends on above.

¹ <https://github.com/nextstrain/cli/pull/357>
² <https://github.com/nextstrain/cli/pull/361>