Check if Rosetta 2 is enabled in the Conda and Docker runner checks
When on macOS on aarch64 hardware, the Conda runtime currently requires
Rosetta.  The Docker runtime does not require it, but can benefit from it
because our aarch64 image isn't totally free of x86_64 code (yet?),
so emulation is required and Rosetta is faster than QEMU.

The standalone installer still checks for Rosetta as we previously
didn't produce aarch64 standalone binaries¹, but that check will be made
conditional on Nextstrain CLI version in a subsequent commit² after the
first release of aarch64 support and these new Rosetta checks.

¹ <https://github.com/nextstrain/cli/pull/357>
² <https://github.com/nextstrain/cli/pull/358>