remote.nextstrain_dot_org: Add failing doctests for destination sidecar suffix handling
Mixed subresource extensions work due to the previous bug fix in
"remote.nextstrain_dot_org: Fix download dots-in-destination bug"
(5d9533f) but empty sidecar suffixes on non-primary subresources do not.

I noted this edge case when working on other fixes in this code.  We
don't currently run into this edge case, but it's good to consider and
prevent anyway to avoid future surprises if we ever do.  Our data model
makes it possible and it's not an unreasonable use case.

The first test fails with:

    Expected:
        ['baz.md', 'baz.json', 'baz_root-sequence.json']
    Got:
        ['baz.md', 'baz_.json', 'baz_root-sequence.json']

and the subsequent similar tests fail similarly.