dev: Upgrade setup-miniconda action from v2 → v3
The latest version will now use Conda's osx-arm64 channel on macOS when
available.  While the original reason for this change was because we
were encountering issues with osx-64 packages on arm64 hardware¹, this
upgrade alone wasn't enough to fix our issues.²  Still, let's upgrade
since v3 is the supported version, and it'll be good to use non-emulated
osx-arm64 binaries going forward anyway.

¹ <https://github.com/nextstrain/cli/issues/382>
² <https://github.com/nextstrain/cli/pull/384#issuecomment-2257479625>