devel/pyoxidizer: Correctly locate just one already-downloaded PyOxidizer binary
Taking the first line never worked as intended, as the input from echo
was never multi-line but space-separated words.  orz

Reverse version sort for good measure to use the newest version instead
of oldest if we have more than one (which was how I ran into this bug in
the first place).