dev: Update GitHub runners used by CI and periodic installer tests
macOS 12 was deprecated in August and will be removed by December.¹  I
moved the build and test jobs to the next oldest version, macOS 13.

Ubuntu 24.04 was released in late September,² so test on it.

macOS 15 was released for beta in late September,² so test on it too.

Test the x86_64 binary on macOS 14 and 15 too since it should work with
Rosetta.  I'm not sure why it wasn't tested on macOS 14 previously;
maybe just an oversight?

¹ <https://github.blog/changelog/2024-08-19-notice-of-upcoming-deprecations-and-breaking-changes-in-github-actions-runners/>
² <https://github.blog/changelog/2024-09-25-actions-new-images-and-ubuntu-latest-changes/>