runner.aws_batch: Gracefully handle errors when fetching logs for completed jobs
For completed jobs, it's more useful to continue on with printing the
job status (e.g. success or reason for failure) and downloading job
results even if an error occurs when fetching logs.  As a concrete
example, we've observed cases where a failed job has a log stream
associated with it in Batch but that log stream does not actually exist
in CloudWatch Logs.¹  The log fetch error hid the reason for job
failure, hampering troubleshooting.

¹ <https://bedfordlab.slack.com/archives/C01LCTT7JNN/p1730406138009409>