util: Move the Conda runner's lax parse_version() into util as parse_version_lax()
The lax implementation is always successful at producing a Version
object suitable for comparisons, even if the version isn't a valid
Python package version.  This makes it useful for non-Python ecosystems,
such as Conda packages, and is what I want to use for pathogen versions
too.

Renames the strict parse_version() previously present in util to
parse_version_strict() because it raises an exception on invalid
versions.  It is still only used internally in util to parse versions
that we expect to be valid.