Handle case where sockaddr is (int, bytes)
Types for the standard library were updated with the possibility of
having an (int, bytes) value for the socket address returned from
getaddrinfo in the case that Python was compiled with --disable-ipv6. An
IP and port combo cannot be resolved from such raw data, so we fall back
to the input host and port.

The type signature is not specific about this happening on IPv6 only, so
we must ignore type checking on the return value.

Co-authored-by: Thomas Sibley <tsibley@fredhutch.org>