Ignore type error from getnameinfo
Types for the standard library were updated with the possibility of
having an (int, bytes) value for the socket address returned from
getaddrinfo in the case that Python was compiled with --disable-ipv6.

It makes sense for getnameinfo to take as input whatever is output from
its inverse function, getaddrinfo, but its type signature has not been
updated. Ignore for now.