Workflows-as-programs: `nextstrain run` and co.
Adds a new command, `nextstrain run`, to run (compatible) pathogen
workflows in a more managed way with easier update paths, without the
need for user-facing Git, with support for multiple versions, and with
support for concurrent-but-separate analyses via the same workflow.

Supported by changes to

  - `nextstrain setup` to obtain and set up specific versions of pathogens
  - `nextstrain update` to keep pathogens up-to-date
  - `nextstrain version` to report on pathogen versions available locally

At the moment, the only compatible pathogen is measles at my
not-yet-finished demo/prototype branch.¹  Avian flu should not be far
behind, though.

There's a lot of functionality (and polish) here and elsewhere still
todo to fully realize the sweeping goals of workflows-as-programs², but
this is a fully-usable first piece of the puzzle that can stand on its
own for now.

¹ <https://github.com/nextstrain/measles/pull/55>
² <https://github.com/nextstrain/public/issues/1>