dev: Fix open_browser() tests to really, truly, actually not pop a browser locally
The previous approach of late-loading of nextstrain.cli.command.view
only worked when running an explicit set of tests/*.py files—pytest's
discovery process for doctests and pytest_*() functions means it loads
all of nextstrain.cli too early otherwise—and most of the time a new tab
or two was popped open because the environment modification was too
late.  This was annoying.  The previous approach also predated our
devel/pytest wrapper to reliably set env by insisting on a common
entrypoint to tests.  With that now present, the easiest and more
reliable thing is to move the setting of BROWSER there before pytest is
ever exec-ed.