Force fresh setup of osx-arm64 runtime version on update
This will automatically switch Apple silicon users from the osx-64
version of the runtime (which requires emulation on such devices) to the
osx-arm64 version.

Note that this does not force a setup if the latest spec version is
already installed.

VERSION_FEATURE and version_supports are based on the similarly named
IMAGE_FEATURE and image_supports in the docker runner code.