runner.conda: Configure Micromamba to use the same CA certificates as requests
This makes it easier to configure alternative CAs in development/testing
in a way that's consistent across components.

By default Micromamba (1.5.8) uses the system CA certificates, the
sourcing of which varies depending on TLS backend.  Micromamba (or
really, libmamba) attempts to respect REQUESTS_CA_BUNDLE explicitly
itself, but the code appears to contain a logic error that means it's
bypassed.  And it does not try to fall back to respecting CURL_CA_BUNDLE
as requests does.  The documentation also says that --cacert-path may be
a directory, but the code supports only files.