runner.conda: Start using osx-arm64 on supported platforms
When supported, new setups will get osx-arm64 instead of osx-64 and
existing setups will be upgraded from osx-64 → osx-arm64.  If the
version of nextstrain-base that's being installed/upgraded (or more
likely, downgraded to) to does not have an osx-arm64 build, then an
osx-64 build will be used instead (provided Rosetta 2 is enabled).

Co-authored-by: Victor Lin <13424970+victorlin@users.noreply.github.com>
Related-to: <https://github.com/nextstrain/cli/pull/379>
Related-to: <https://github.com/nextstrain/conda-base/pull/80>