runner.conda: Drop sorting by build_number and note why we ignore build variants
The sorting by build number originally came about when we were still
using the package_distribution() function to find the Micromamba dist to
download.  As Micromamba had used "fixup" builds in the past, it seemed
useful/important to include build number in the sort.  However, builds
can also be used for "variants", e.g. differing dependencies, and
resolving which "variant" build to use requires the solver not a simple
sort.  We don't need to support the way Micromamba used/uses builds
anymore, and since our usage of multiple builds for nextstrain-base is
likely to be for variants, we can't properly decide which to use anyway.