runner.conda: Stop checking Content-Type for Micromamba download
Just try to open it as a tar archive (possibly compressed) and handle
the error if it's not readable.  Now that we're not always fetching from
a single expected source (i.e. Anaconda), this avoids issues with
different mirrors reporting different Content-Types (not to mention
responses which claim to an expected Content-Type but aren't actually).

We may still want to use Content-Type to dispatch to different format
handlers in the future, but right now we only handle tar.

Motivated by @victorlin's review.¹

¹ <https://github.com/nextstrain/cli/pull/436#discussion_r2091882960>