runner.conda: Allow overriding of subdir selection
It seems like it will prove useful to be able to force a particular
subdir for both testing/development and as a workaround to issues with
specific subdirs on platforms that support more than one (e.g. forcing
osx-64 instead of osx-arm64, if osx-arm64 is broken in some way).

I thought about implementing this along with my recent subdir handling
changes, but I thought maybe YAGNI and deferred.  The slew of issues
recently uncovered with osx-arm64 though made me reconsider and JFDI.