dev: Fix Pyright failures with latest version (1.1.401)
An included typeshed stub update¹ for docutils (and many other modules)
meant that docutils.readers.standalone.Reader became an unknown type and
was duly reported as such.  We don't care.

An apparent regression² results in unnecessary "# type: …" comments in
the typeshed itself's internal code being reported as our problem.  It
is not.  Ignore them.

Note that in this case of ignoring out-of-project source, matching an
absolute path seems to be required for the Pyright "ignore" pattern to
actually work.  (I found no relative or partial paths that work.)  Also
note that whatever globbing Pyright is doing will not match a leading
"**" to an absolute path.  Give up and anchor the pattern to an absolute
root separately for POSIX and Windows.  Yay.

¹ <https://github.com/python/typeshed/issues/13765>
  <https://github.com/python/typeshed/pull/14030>

² <https://github.com/microsoft/pyright/issues/10487>