docs: Include a changelog in the documentation
This has been a longstanding omission that I finally got around to
correcting because I wanted to more publicly link to the changelog and
not point people into the den of source-ry that is GitHub.

I did this thing not because it was easy, but because I *thought* it
would be easy: just drop in CHANGES.md as-is using an "include"
directive.  Well, after much gnashing of keys, it turns out that usable
heading anchors are impossible to achieve with this method unless we
make CHANGES.md include noisy MyST syntax.  So I swapped to doc
generation/transformation instead.

Fixes broken links (and removes one that's gone) in the changelog now
that they're caught by the linkchecker.