Don't always recommend creating a new compute environment
I was able to set this up successfully with an existing compute
environment. My understanding, based on the message which is now in a
note, is that creating a new compute environment is only necessary when
the existing environment was created with the launch template version
set to `$Latest`, which won't happen for users following this guide from
scratch.