Ignore environment settings in pathogen repo request
The default behavior of the requests library is to automatically honor¹
the user's netrc file. We do not (1) need or (2) want this behavior in
the case of downloading a pathogen repo because (1) the API endpoint
works fine without authentication and (2) any expired credentials will
cause the request to fail when it would've succeeded without the
credentials.

I chose not to set trust_env to False on all Sessions because the
default honors² REQUESTS_CA_BUNDLE and CURL_CA_BUNDLE env vars, a
feature used³ by other requests.

¹ <https://github.com/psf/requests/blob/v2.32.4/README.md#supported-features--bestpractices>
² <https://github.com/psf/requests/blob/v2.32.4/src/requests/sessions.py#L768-L769>
³ <https://github.com/nextstrain/docs.nextstrain.org/blob/ab09ba2577b0333c42dbff2e5b293013c226ab9a/src/reference/ca-certificates.rst#requests-module>