Document repo rulesets
I was unaware of these rulesets, so the following output when pushing a
release tag was concerning:

    remote: Resolving deltas: 100% (6/6), completed with 3 local objects.
    remote: Bypassed rule violations for refs/tags/10.2.1:
    remote:
    remote: - Cannot create ref due to creations being restricted.
    remote:

Only after inspecting the repo rulesets did I realize that the output
can be disregarded. Hopefully the added snippet will help others come to
the same realization without needing to inspect repo rulesets.