dev: Properly stage changes to MyST-formatted changelog
The previous commit "dev: Update MyST-formatted changelog during
release" (e360cfff) added the command to regenerate the MyST-formatted
changelog, however it did not stage the changes in the same fashion as
the other changelog file, so the generated commits ended up being out of
sync. I've tested locally to check that it's fixed now.