pathogens: support slashes in version name
Previously, if the version included a slash, e.g. a branch name `foo/bar`,
the request to the Nextstrain server would fail with a 404 error even if the 
branch existed. This is because the slashes in the version were not properly 
escaped in the URL and therefore was not treated as a single route parameter.
Note I did not do the same for the pathogen name since GitHub repo names cannot 
contain slashes.

Resolves https://github.com/nextstrain/cli/issues/458