runner.aws_batch: Download .snakemake/storage/ too
When using Snakemakes storage support to download remote files, Snakemake stores 
the downloaded files in `.snakemake/storage` by default. This allows users to 
keep the remote files that were downloaded during the build on AWS Batch.

Note that the storage path is configurable via Snakemake's 
`--local-storage-prefix`.¹ So if someone configures the storage path to a custom 
path within `.snakemake`, e.g. `.snakemake/foo`, then it would not be available 
in their downloaded workdir. I'm not even sure it's possible to configure the 
path in entrypoint-aws-batch for the docker-base image, so I've added a warning 
against using the Snakemake option if using the aws-batch runtime.

Resolves https://github.com/nextstrain/cli/issues/453
Related to https://github.com/nextstrain/docker-base/pull/259

¹ <https://github.com/snakemake/snakemake/blob/v9.6.3/src/snakemake/cli.py#L1463-L1469>