util: Account for non-default standalone installation paths in suggested update command
Sets DESTINATION to the installation path we're running under so the new
version is installed to the same place, rather than the default path.

This muddies the commands but makes them behave better, which is more
important.  Particularly since soon we'll offer to run these commands
automatically for the user.