Add `nextstrain update cli` to perform a self-upgrade
Provides a consistent upgrade command regardless of installation method.

Although "self" is a term that's used by other programs for this type of
update (e.g. "self-upgrade", "update self") and it's come up in our
discussions of this feature, using the term "cli" seems clearer and a
more consistent identity for the `nextstrain` program.

Messages pertaining to the CLI update have their terminology switched
from "upgrade" to "update" for consistency with the command name and
other runtime and pathogen updates.  And the subject of the updates
("Nextstrain CLI") is added to more places in those messages for
clarity.

I worry the commands we run won't always work right, but presumably
we'll hear about it and can fix issues that crop up.

Relatedly, `nextstrain update` could really use a --dry-run mode and
support would not be too difficult to add.  (It's already there in the
Docker and Singularity runtimes.)  But I'm leaving that for another time
and letting folks rely on `nextstrain check-setup` to check for/preview
any CLI update.

Resolves: <https://github.com/nextstrain/cli/issues/269>