Hint at invalid credentials when handling 401/403 errors during pathogen download
Since netrc credentials are automatically and silently added, they may
be unknown to the user.  Note their presence and suggest solutions in a
hint following the error.  We do not automatically retry without
credentials because they may be intended by (and known to) the user.

Ideally this sort of error handling is implemented at a higher (e.g.
within nextstrain.cli.run()) or lower (e.g. within
nextstrain.cli.requests.Session) level in the call stack so that ~all
requests benefit from it, not just the single request targeted here.
I'm leaving that for future work, though, as a) that's a more extensive
change and b) we expect this specific request to be more affected by
stale netrc credentials due to the combination of requests for
api.github.com and old versions of the gh CLI.

Resolves: <https://github.com/nextstrain/cli/issues/444>
Related-to: <https://github.com/nextstrain/cli/pull/478>