version: Declutter pathogen workflow listing
List the workflow names on the same line as the pathogen name@version
when there are any workflows and emit nothing when there aren't.  This
fits more in the style of the rest of `nextstrain version` and is much
less cluttered (more information dense), particularly considering that
we, in short order too, expect ~all of our pathogens (and all their
future versions) to have the same 2–3 workflows listed (nextclade,
ingest, phylogenetic).

Prompted by feedback from @tsibley.¹

¹ <https://github.com/nextstrain/cli/pull/472#discussion_r2373235440>

Co-authored-by: Thomas Sibley <tsibley@fredhutch.org>