pathogens: Expand 401/403 error message to pinpoint the netrc file path
This is more useful to someone encountering the error, rather than a)
not mentioning any paths or b) mentioning multiple paths and making the
user figure out which one is responsible.  Granted, it's a bit more code
for us, but "don't make a person do a computer's work" is a pretty good
principle.

Prompted in part by @victorlin's review comments.¹

¹ <https://github.com/nextstrain/cli/pull/479#discussion_r2379863664>